# Flatten json