--- a/CHANGES.md
+++ b/CHANGES.md
@@ -17,6 +17,7 @@ Note that the top-most release is changes in the unreleased master branch on Git
 - `Arche.glance()`, #88
 - Item links in Schema validation errors, #89
 - Empty NAN bars on category graphs, #93
+- `expand=True` which enables nested data flattening is more than 100x faster and consumes ~2x less memory than before, #94
 ### Removed