--- a/setup.cfg
+++ b/setup.cfg
@@ -21,6 +21,7 @@ packages = find:
 setup_requires = setuptools_scm
 install_requires = 
     pandas
+    flatten_json
     scrapinghub[msgpack]
     plotly>=3.8.0
     genson