… src/arche/tools/pandas.py
… tests/tools/test_pandas.py