This one significantly increases performance and decreases memory usage if using `flat_df`, which corresponds to `expand=True` default parameter.
Some things to consider:
1. If original column name and expanded column name the same, the original column will be dropped. 
See here https://github.com/amirziai/flatten/issues/48
I think it won't happen in 99% of cases, I've never encountered such a case. Alternatively, we can set other separator which potentially bumps it to 99.9%.
Nevertheless, there's is that issue and also a comment in tests.

Time and peak memory, 1k samples from a heavily nested df (see `shared/qa/Experiments/data/items_epicsports_150k_nested.pickle`).
This pr - 1.63s, 2400 MB
Old - 4min 57s, 4411.51 MiB
#74 - 6min 2s, 2800 MB

And, interestingly, all 150k items with this pr:
9min 40s
peak memory: 3682.77 MiB, increment: 1391.34 MiB